refactor put onboardingApi into userApi
