+++ package.json
@@ -35,6 +35,7 @@
     "@mui/x-date-pickers": "^5.0.3",
     "@reduxjs/toolkit": "^1.8.5",
     "date-fns": "^2.29.3",
+    "isomorphic-fetch": "^3.0.0",
     "next": "12.3.0",
     "notistack": "^2.0.5",
     "react": "18.2.0",
@@ -57,6 +58,7 @@
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
+    "@types/isomorphic-fetch": "^0.0.36",
     "@types/jest": "^29.0.1",
     "@types/node": "18.7.17",
     "@types/react": "^18.0.19",
