+++ src/core/services/User/userApi.ts
@@ -1,29 +1,60 @@
 import { authQuery } from "../authQuery";
 import { createApi } from "@reduxjs/toolkit/query/react";
-import type { IUser, IUserAccount } from "@/core/store/User";
+// import { RTKTags } from "@/core/store";
+import type { IUser, IUserAccount } from "./userApiTypes";
+import type { IOnboardingInfoFormSchema } from "@/modules/Onboarding/components";
 
 export const userApi = createApi({
   reducerPath: "userApi",
   baseQuery: authQuery("user"),
+  tagTypes: ["User"],
   endpoints: (builder) => ({
+    getUser: builder.query<IUser, void>({
+      query: () => ({
+        url: "/",
+        method: "GET",
+      }),
+      providesTags: ["User"],
+    }),
     getUserAccount: builder.query<{ account: IUserAccount }, void>({
       query: () => ({
         url: "account",
         method: "GET",
         credentials: "include",
       }),
+      providesTags: ["User"],
     }),
-    getUser: builder.query<IUser, void>({
-      query: () => ({
-        url: "/",
-        method: "GET",
+    onboardingUpdateInfo: builder.mutation<void, IOnboardingInfoFormSchema>({
+      query: (info) => ({
+        url: "onboarding/info",
+        method: "PUT",
+        body: info,
       }),
+      invalidatesTags: ["User"],
+    }),
+    onboardingUpdateGender: builder.mutation<void, { gender: string }>({
+      query: (body) => ({
+        url: "onboarding/gender",
+        method: "PUT",
+        body: body,
+      }),
+      invalidatesTags: ["User"],
     }),
   }),
 });
 
-export const { useGetUserAccountQuery, useGetUserQuery } = userApi;
+export const {
+  useGetUserAccountQuery,
+  useGetUserQuery,
+  useOnboardingUpdateGenderMutation,
+  useOnboardingUpdateInfoMutation,
+} = userApi;
 
 export const {
-  endpoints: { getUserAccount, getUser },
+  endpoints: {
+    getUserAccount,
+    getUser,
+    onboardingUpdateGender,
+    onboardingUpdateInfo,
+  },
 } = userApi;
