+++ src/core/services/index.ts
@@ -1,1 +1,4 @@
+export * from "./User/userApi";
+export * from "./User/userApiTypes";
+export * from "./authQuery";
 export * from "./baseQuery";
