--- src/core/store/User/index.ts
@@ -1,2 +1,0 @@
-export * from "./userSlice";
-export * from "./userTypes";
