+++ src/core/store/index.ts
@@ -1,2 +1,4 @@
+export * from "./middlewares";
 export { default as reducers } from "./reducers";
+export { default as storage } from "./storage";
 export * from "./store";
