+++ test/setupTests.ts
@@ -1,3 +1,4 @@
+import "isomorphic-fetch";
 import "@testing-library/jest-dom";
 import { loadEnvConfig } from "@next/env";
